+++ CHANGELOG.md
@@ -24,7 +24,7 @@
 - relay: add infos "relay_api_version" and "relay_api_version_number"
 - core: add hdata count in evaluation of expressions with `hdata_count:name[list]` or `hdata_count:name[pointer]`
 - core: add completion "bars_items"
-- api, relay: send new signal "buffer_line_data_changed" when a line is updated in a buffer via hdata, send event "buffer_line_data_changed" to API clients
+- api, relay: send new signal "buffer_line_data_changed" when a line is updated in a buffer via hdata, send event "buffer_line_data_changed" to clients of "api" and "weechat" protocols
 - api: add hashtable type "longlong"
 - api: add function line_search_by_id
 - doc: add doc on "api" relay